refactor: blog data を json から読み込むように修正
